+++ src/main/java/it/contrader/model/Conversion.java
@@ -31,10 +31,10 @@ public class Conversion {
 	private Long idConversion;
 	
 	@Column(unique = false)
-	private int idUser;
+	private Long idUser;
 	private String source;
 	private SourceType sourceType;
 	private OutputType outputType;
-	private int changes;
+	private Long changes;
 	
 }
